refactor: improve separation of concerns
create validation and auth utilz and better separate what controller vs
useCase layers should do

+++ src/app/api/users/useCase/remove.ts
@@ -1,7 +1,13 @@
-import { NotFoundError } from "@/app/api/errors";
+import { NotFoundError, UnauthorizedError } from "@/app/api/errors";
 import { UserRepository } from "@/app/api/users/repository";
+import util from "@/app/api/utils";
 
-export async function remove(id: string) {
+export async function remove(id: string, sessionId: string) {
+  const session = await util.validateSession(sessionId);
+
+  if (session?.userId !== id)
+    throw new UnauthorizedError("Cannot delete other user's account");
+
   const repository = new UserRepository();
   const user = await repository.remove(id);
